add dom updates
