+++ src/domUpdates.js
@@ -1,4 +1,6 @@
 import {currentUser} from './scripts.js'
+import RecipeRepository from './classes/RecipeRepository.js';
+import {recipeData} from './scripts.js'
 
 let domUpdates = {
     hide(element){
@@ -35,7 +37,7 @@ let domUpdates = {
     },
     showHomeView() {
         domUpdates.show(recipeGrid);
-        generateRandomHomeViewRecipes();
+        scripts.generateRandomHomeViewRecipes();
         domUpdates.hide(addRecipeForm);
         domUpdates.hide(allRecipeContainer);
         domUpdates.hide(allRecipeGrid);
@@ -61,10 +63,7 @@ let domUpdates = {
           if (currentUser.recipesToCook.includes(recipe)){
             buttonClassesToCook = "recipesToCook is-saved"
           }
-        //   if(currentUser.recipesToCook.length===0){
-        //     allRecipeGrid.innerHTML = `<h1>No recipes found!</h1>`
-        //     return
-        //   }
+
           allRecipeGrid.innerHTML +=
             `<article class="mini-recipe" id="${recipe.id}">
              <img src= "${recipe.image}" alt= "${recipe.name}">
@@ -87,6 +86,51 @@ let domUpdates = {
             popupContainer.classList.remove(`${color}-popup`)
             domUpdates.hide(popupContainer);
         }, timeInMS)
+},
+
+  displayEmptyFavorites() {
+    featuredRecipes.innerHTML = `<h1>Add a Recipe</h1>`;
+},
+
+displaySavedRecipes() {
+  if(currentUser.favoriteRecipes.length===0){
+    allRecipeGrid.innerHTML = `<h1>No recipes found!</h1>`
+    return
+  }
+  featuredRecipes.innerHTML = `<h1>Saved Recipes</h1>`;
+},
+
+generateRandomHomeViewRecipes(){
+domUpdates.show(featuredRecipes);
+let recipeRepo = new RecipeRepository(recipeData);
+let randomRecipeIndex1 = Math.floor(Math.random() * recipeRepo.recipeData.length)
+let randomRecipeIndex2 = Math.floor(Math.random() * recipeRepo.recipeData.length)
+let randomRecipeIndex3 = Math.floor(Math.random() * recipeRepo.recipeData.length)
+if (randomRecipeIndex2 === randomRecipeIndex1){
+  randomRecipeIndex2 = Math.floor(Math.random() * recipeRepo.recipeData.length)
+}
+if (randomRecipeIndex3 === randomRecipeIndex1 || randomRecipeIndex2){
+  randomRecipeIndex3 = Math.floor(Math.random() * recipeRepo.recipeData.length)
+}
+recipeGrid.innerHTML = ""
+let randomRecipesIndex = [randomRecipeIndex1, randomRecipeIndex2, randomRecipeIndex3]
+randomRecipesIndex.forEach((randomRecipeIndex) => {
+
+  let index = randomRecipeIndex
+
+ recipeGrid.innerHTML +=
+  `<article class="recipe" id=${recipeRepo.recipeData[index].id}>
+  <div class="meal-image">
+    <img src="${recipeRepo.recipeData[index].image}" alt="meal image" class="image">
+  </div>
+  <div class="recipe-content">
+    <button type="favoriteStar" name="favoriteStar" class="favorite-star" id="favoriteStar1">♡</button>
+    <div class="recipe-info">
+      <h2>${recipeRepo.recipeData[index].name}</h2>
+    </div>
+  </div>
+</article>`
+})
 },
 
 }
